requirements.txt
asgiref==3.8.1
Brotli==1.1.0
build==1.0.3
CacheControl==0.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
class-registry==2.1.2
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-jsonfield==1.4.1
django-perms==2.1.0
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.12.2
filters==1.3.2
filters-django==1.0.5
httpie==3.2.2
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.5
packaging==23.2
pexpect==4.8.0
pillow==10.3.0
pipenv==2023.12.1
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.1
rapidfuzz==3.5.2
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
SecretStorage==3.3.3
sendgrid==6.11.0
shellingham==1.5.3
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing==3.7.4.3
typing_extensions==4.12.0
urllib3==2.0.6
validation==0.8.3
virtualenv==20.25.1
whitenoise==6.6.0
zipp==3.17.0
